Split cli/commands/projects.ts into read/mutate/maintenance modules
