+++ src/cli/commands/projects.ts
@@ -1,48 +1,29 @@
-import { promises as fs } from 'node:fs';
-import { dirname, join, relative, resolve, sep } from 'node:path';
-import { findProjectReadmes } from '../../main/walk';
-import { parseReadme } from '../../main/parse';
-import { appendTimelineEntry, parseTimelineEntries, transitionStatus } from '../../main/mutate';
-import { search as searchAll } from '../../main/search';
-import { regenerateIndex, type IndexRegenReport } from '../../main/index-tree';
-import { projectsStrategy } from '../../main/index-projects';
-import { checkBranchState } from '../../main/worktree-ops';
-import { exec } from '../../main/exec';
-import { touchDirtyMarker } from '../../main/dirty';
+import { CliError, ExitCodes, type OutputContext } from '../output';
+import { type ParsedArgs } from '../parser';
 import {
+  listProjects,
+  readProject,
+  resolveCommand,
+  searchProjects,
+  validateCommand,
+} from './projects-read';
+import { statusCommand, closeProject, reopenProject } from './projects-mutate';
+import {
+  indexCommand,
+  createCommand,
+  scanPromotionsCommand,
+  rewriteHeadersCommand,
+  backfillClosed,
   createProjectCore,
-  type CreateProjectInput,
-  type CreateProjectResult,
-} from '../../main/create-project';
-import { rewriteHeadersInTree, type RewriteHeadersReport } from '../../main/rewrite-headers';
-import { KNOWN_STATUSES, type SearchHit } from '../../shared/types';
-import { statusOrder } from '../../shared/projects';
-import { isValidSlugTail } from '../../shared/slug';
-import { resolveSlug } from '../slug-resolver';
-import { CliError, ExitCodes, emit, validation, type OutputContext } from '../output';
-import { parseHeader, validateHeader, validateBody, type HeaderFields } from '../../shared/header';
-import { readHeader } from '../../main/header-io';
-import { assertNoExtraFlags, parseCsvFlag, parseIntFlag, type ParsedArgs } from '../parser';
+  isValidSlugTail,
+} from './projects-maintenance';
+import type { CreateProjectInput, CreateProjectResult } from './projects-maintenance';
 
-const PROMOTION_RE =
-  /(^|\b)(always|never|must|convention|rule|pattern|whenever|all (apps|sites|projects))\b/i;
-
-interface ProjectListRow {
-  slug: string;
-  path: string;
-  absPath: string;
-  title: string;
-  kind: string;
-  status: string;
-  apps: string[];
-  branch: string | null;
-  base: string | null;
-  date: string | null;
-  closedAt: string | null;
-  stepCounts: { todo: number; doing: number; done: number; dropped: number };
-  deliverableCount: number;
-  headerWarnings: { field: string; message: string }[];
-}
+// Re-exports kept on the historical import path. `createProjectCore` and
+// helpers live in src/main/create-project.ts; we surface them here so any
+// external consumer that imported from this file keeps working.
+export { createProjectCore, isValidSlugTail };
+export type { CreateProjectInput, CreateProjectResult };
 
 export async function runProjects(
   verb: string | null,
@@ -81,880 +62,6 @@ export async function runProjects(
     default:
       throw new CliError(ExitCodes.USAGE, `Unknown projects verb: ${verb}`);
   }
-}
-
-async function indexCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const dryRun = args.flags['dry-run'] === true;
-  const rewriteAggregated = args.flags['rewrite-aggregated'] === true;
-  delete args.flags['dry-run'];
-  delete args.flags['rewrite-aggregated'];
-  assertNoExtraFlags(args);
-  const report = await regenerateIndex(conceptionPath, projectsStrategy, {
-    dryRun,
-    rewriteAggregated,
-  });
-  emit(ctx, report, formatIndexReport, [], { streamField: 'updated' });
-}
-
-function formatIndexReport(report: IndexRegenReport): string {
-  const lines: string[] = [];
-  lines.push(
-    `Index regeneration: ${report.tree}/  (${report.dryRun ? 'dry-run' : 'wrote changes'})`,
-  );
-  if (report.created.length > 0) {
-    lines.push(`Created (${report.created.length}):`);
-    for (const p of report.created) lines.push(`  + ${p}`);
-  }
-  if (report.updated.length > 0) {
-    lines.push(`Updated (${report.updated.length}):`);
-    for (const u of report.updated) {
-      const parts: string[] = [];
-      if (u.added.length > 0) parts.push(`added ${u.added.length}`);
-      if (u.dropped.length > 0) parts.push(`dropped ${u.dropped.length}`);
-      if (u.tagsAdded.length > 0) parts.push(`tags+ ${u.tagsAdded.length}`);
-      lines.push(`  ~ ${u.indexPath}  (${parts.join(', ')})`);
-    }
-  }
-  if (report.unchanged.length > 0) {
-    lines.push(`Unchanged: ${report.unchanged.length}`);
-  }
-  if (report.flaggedRenames.length > 0) {
-    lines.push(`Suspected renames (${report.flaggedRenames.length}):`);
-    for (const r of report.flaggedRenames) {
-      lines.push(`  ? ${r.indexPath}  ${r.oldName}  →  ${r.newName}`);
-    }
-  }
-  if (report.overTagDropped.length > 0) {
-    lines.push(`Cap reached, dropped surplus tags (${report.overTagDropped.length}):`);
-    for (const o of report.overTagDropped) {
-      lines.push(`  · ${o.indexPath}  ${o.entry}  dropped: ${o.dropped.join(', ')}`);
-    }
-  }
-  if (report.validationWarnings.length > 0) {
-    lines.push(`Item validation (${report.validationWarnings.length}):`);
-    for (const w of report.validationWarnings) {
-      lines.push(
-        `  [${w.severity}] ${relativeIfPossible(w.path, report.rootPath)}  ${w.field}: ${w.message}`,
-      );
-    }
-  }
-  if (report.dirtyClear) lines.push('Dirty marker cleared.');
-  return lines.join('\n') + '\n';
-}
-
-function relativeIfPossible(path: string, rootPath: string): string {
-  // rootPath is .../projects or .../knowledge; show paths relative to its
-  // parent (the conception root) when possible. Falls back to the absolute
-  // path when the file lives outside the root (e.g. `--path` overrides).
-  if (!path.startsWith(rootPath)) return path;
-  const conceptionRoot = dirname(rootPath);
-  return relative(conceptionRoot, path) || path;
-}
-
-async function createCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const apps = parseCsvFlag(args.flags.apps) ?? [];
-  if (apps.length === 0) validation(`--apps is required (comma-separated, may be backticked)`);
-  const input = {
-    kind: String(args.flags.kind ?? '').toLowerCase(),
-    slug: String(args.flags.slug ?? '').trim(),
-    title: String(args.flags.title ?? '').trim(),
-    apps,
-    branch: typeof args.flags.branch === 'string' ? args.flags.branch.trim() || null : null,
-    base: typeof args.flags.base === 'string' ? args.flags.base.trim() || null : null,
-    date: typeof args.flags.date === 'string' ? args.flags.date.trim() : undefined,
-    severity:
-      typeof args.flags.severity === 'string' ? args.flags.severity.toLowerCase() || null : null,
-    severityImpact:
-      typeof args.flags['severity-impact'] === 'string'
-        ? String(args.flags['severity-impact']).trim() || null
-        : null,
-    environment:
-      typeof args.flags.environment === 'string'
-        ? args.flags.environment.trim().toUpperCase() || null
-        : null,
-  };
-  for (const k of [
-    'apps',
-    'kind',
-    'slug',
-    'title',
-    'branch',
-    'base',
-    'date',
-    'severity',
-    'severity-impact',
-    'environment',
-  ]) {
-    delete args.flags[k];
-  }
-  assertNoExtraFlags(args);
-  const result = await createProjectCore(conceptionPath, input);
-  emit(ctx, result, (data) => {
-    const d = data as { relPath: string; readme: string };
-    return `Created ${d.relPath}\n  README: ${d.readme}\n`;
-  });
-}
-
-// createProjectCore + CreateProjectInput + CreateProjectResult moved to
-// src/main/create-project.ts in pass-10 — re-exported here so external
-// consumers (if any) keep the historical import path.
-export { createProjectCore };
-export type { CreateProjectInput, CreateProjectResult };
-export { isValidSlugTail };
-
-async function scanPromotionsCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const slug = args.positional[0];
-  if (!slug) {
-    throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects scan-promotions <slug>');
-  }
-  assertNoExtraFlags(args);
-  const candidate = await resolveSlug(conceptionPath, slug);
-  const notesDir = join(candidate.itemDir, 'notes');
-  let entries: string[];
-  try {
-    entries = (await fs.readdir(notesDir)).filter((n) => n.toLowerCase().endsWith('.md')).sort();
-  } catch (err) {
-    if ((err as NodeJS.ErrnoException).code === 'ENOENT') entries = [];
-    else throw err;
-  }
-  const candidates: { relPath: string; line: number; paragraph: string; match: string }[] = [];
-  for (const name of entries) {
-    const abs = join(notesDir, name);
-    const raw = await fs.readFile(abs, 'utf8');
-    const lines = raw.split(/\r?\n/);
-    for (let i = 0; i < lines.length; i++) {
-      const m = PROMOTION_RE.exec(lines[i]);
-      if (!m) continue;
-      const paragraph = extractParagraph(lines, i);
-      candidates.push({
-        relPath: `notes/${name}`,
-        line: i + 1,
-        paragraph,
-        match: m[0],
-      });
-    }
-  }
-  emit(
-    ctx,
-    {
-      slug: candidate.slug,
-      path: candidate.itemDir,
-      candidates,
-    },
-    (data) => {
-      const d = data as typeof data & { candidates: typeof candidates };
-      if (d.candidates.length === 0) return '(no promotion candidates found)\n';
-      const out: string[] = [];
-      for (const c of d.candidates) {
-        out.push(`${c.relPath}:${c.line}  [${c.match}]`);
-        for (const para of c.paragraph.split('\n')) out.push(`  ${para}`);
-      }
-      return out.join('\n') + '\n';
-    },
-  );
-}
-
-function extractParagraph(lines: string[], hitIndex: number): string {
-  let start = hitIndex;
-  while (start > 0 && lines[start - 1].trim() !== '') start--;
-  let end = hitIndex;
-  while (end < lines.length - 1 && lines[end + 1].trim() !== '') end++;
-  return lines
-    .slice(start, end + 1)
-    .join('\n')
-    .trim();
-}
-
-async function rewriteHeadersCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const dryRun = args.flags['dry-run'] === true;
-  delete args.flags['dry-run'];
-  assertNoExtraFlags(args);
-  const report = await rewriteHeadersInTree(conceptionPath, { dryRun });
-  emit(ctx, { ...report, dryRun }, (data) => {
-    const d = data as RewriteHeadersReport & { dryRun: boolean };
-    const lines: string[] = [];
-    lines.push(`README header rewrite (${d.dryRun ? 'dry-run' : 'wrote changes'})`);
-    if (d.rewritten.length > 0) {
-      lines.push(`Rewritten (${d.rewritten.length}):`);
-      for (const p of d.rewritten) lines.push(`  ~ ${relative(conceptionPath, p)}`);
-    }
-    if (d.alreadyYaml.length > 0) {
-      lines.push(`Already YAML: ${d.alreadyYaml.length}`);
-    }
-    if (d.skipped.length > 0) {
-      lines.push(`Skipped (${d.skipped.length}):`);
-      for (const s of d.skipped) {
-        lines.push(`  ! ${relative(conceptionPath, s.path)}  [${s.reason}]`);
-      }
-    }
-    return lines.join('\n') + '\n';
-  });
-}
-
-async function listProjects(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const statusFilter = parseCsvFlag(args.flags.status);
-  const kindFilter = parseCsvFlag(args.flags.kind);
-  const appsFilter = parseCsvFlag(args.flags.apps);
-  const branchFilter = typeof args.flags.branch === 'string' ? args.flags.branch : null;
-  const sort = (args.flags.sort as string | undefined) ?? 'status';
-  for (const k of ['status', 'kind', 'apps', 'branch', 'sort']) delete args.flags[k];
-  assertNoExtraFlags(args);
-
-  const readmes = await findProjectReadmes(conceptionPath);
-  const rows: ProjectListRow[] = [];
-  for (const readme of readmes) {
-    const project = await parseReadme(readme);
-    const headerFields = parseHeader(await fs.readFile(readme, 'utf8'));
-    const headerWarnings = validateHeader(headerFields, readme).warnings;
-
-    const apps = headerFields.apps;
-    const branch = headerFields.branch;
-
-    if (statusFilter && !statusFilter.includes(project.status)) continue;
-    if (kindFilter && !kindFilter.includes(project.kind)) continue;
-    if (appsFilter && !appsFilter.some((app) => apps.includes(app))) continue;
-    if (branchFilter && branch !== branchFilter) continue;
-
-    rows.push({
-      slug: project.slug,
-      path: relative(conceptionPath, readme.replace(/\/README\.md$/, '')),
-      absPath: readme.replace(/\/README\.md$/, ''),
-      title: project.title,
-      kind: project.kind,
-      status: project.status,
-      apps,
-      branch,
-      base: headerFields.base,
-      date: headerFields.date,
-      closedAt: project.closedAt,
-      stepCounts: project.stepCounts,
-      deliverableCount: project.deliverableCount,
-      headerWarnings,
-    });
-  }
-
-  rows.sort(makeSorter(sort));
-  emit(ctx, rows, formatListHuman);
-}
-
-function makeSorter(sort: string): (a: ProjectListRow, b: ProjectListRow) => number {
-  if (sort === 'slug') return (a, b) => a.slug.localeCompare(b.slug);
-  if (sort === 'date') return (a, b) => (b.date ?? '').localeCompare(a.date ?? '');
-  return (a, b) => {
-    const orderA = statusOrder(a.status);
-    const orderB = statusOrder(b.status);
-    if (orderA !== orderB) return orderA - orderB;
-    return a.slug.localeCompare(b.slug);
-  };
-}
-
-function formatListHuman(rows: ProjectListRow[]): string {
-  if (rows.length === 0) return '(no projects match)\n';
-  const widths = {
-    status: Math.max(6, ...rows.map((r) => r.status.length)),
-    kind: Math.max(4, ...rows.map((r) => r.kind.length)),
-    title: Math.min(60, Math.max(5, ...rows.map((r) => r.title.length))),
-  };
-  const lines: string[] = [];
-  for (const r of rows) {
-    const apps = r.apps.length ? `[${r.apps.join(',')}]` : '';
-    const branch = r.branch ? `(${r.branch})` : '';
-    const title =
-      r.title.length > widths.title ? r.title.slice(0, widths.title - 1) + '…' : r.title;
-    lines.push(
-      `${r.status.padEnd(widths.status)}  ${r.kind.padEnd(widths.kind)}  ${title.padEnd(widths.title)}  ${r.path}  ${apps}  ${branch}`.trimEnd(),
-    );
-  }
-  return lines.join('\n') + '\n';
-}
-
-async function readProject(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const slug = args.positional[0];
-  if (!slug) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects read <slug>');
-  const candidate = await resolveSlug(conceptionPath, slug);
-  const project = await parseReadme(candidate.readmePath);
-  const header = parseHeader(await fs.readFile(candidate.readmePath, 'utf8'));
-  const data: Record<string, unknown> = {
-    slug: candidate.slug,
-    path: candidate.relPath,
-    absPath: candidate.itemDir,
-    title: project.title,
-    kind: project.kind,
-    status: project.status,
-    date: header.date,
-    closedAt: project.closedAt,
-    apps: header.apps,
-    branch: header.branch,
-    base: header.base,
-    summary: project.summary,
-    stepCounts: project.stepCounts,
-    steps: project.steps,
-    deliverables: project.deliverables,
-    deliverableCount: project.deliverableCount,
-    extra: header.extra,
-  };
-  const withNotes = args.flags['with-notes'] === true;
-  delete args.flags['with-notes'];
-  assertNoExtraFlags(args);
-  if (withNotes) {
-    const notesDir = join(candidate.itemDir, 'notes');
-    data.notes = await readNotesDir(notesDir);
-  }
-  emit(ctx, data, formatReadHuman);
-}
-
-async function readNotesDir(dir: string): Promise<{ relPath: string; content: string }[]> {
-  let entries: string[];
-  try {
-    entries = (await fs.readdir(dir)).filter((n) => n.endsWith('.md')).sort();
-  } catch (err) {
-    if ((err as NodeJS.ErrnoException).code === 'ENOENT') return [];
-    throw err;
-  }
-  const out: { relPath: string; content: string }[] = [];
-  for (const name of entries) {
-    const content = await fs.readFile(join(dir, name), 'utf8');
-    out.push({ relPath: `notes/${name}`, content });
-  }
-  return out;
-}
-
-function formatReadHuman(data: Record<string, unknown>): string {
-  const lines: string[] = [];
-  lines.push(`# ${data.title}`);
-  lines.push('');
-  lines.push(`Status: ${data.status}    Kind: ${data.kind}    Date: ${data.date ?? '(none)'}`);
-  if (data.apps && Array.isArray(data.apps) && data.apps.length > 0) {
-    lines.push(`Apps:   ${(data.apps as string[]).join(', ')}`);
-  }
-  if (data.branch) lines.push(`Branch: ${data.branch}`);
-  if (data.base) lines.push(`Base:   ${data.base}`);
-  lines.push(`Path:   ${data.path}`);
-  const counts = data.stepCounts as { todo: number; doing: number; done: number; dropped: number };
-  lines.push(
-    `Steps:  ${counts.todo} todo, ${counts.doing} in-progress, ${counts.done} done, ${counts.dropped} dropped`,
-  );
-  if (data.summary) {
-    lines.push('');
-    lines.push(String(data.summary));
-  }
-  return lines.join('\n') + '\n';
-}
-
-async function resolveCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const slug = args.positional[0];
-  if (!slug) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects resolve <slug>');
-  assertNoExtraFlags(args);
-  const candidate = await resolveSlug(conceptionPath, slug);
-  emit(
-    ctx,
-    {
-      slug: candidate.slug,
-      path: candidate.relPath,
-      absPath: candidate.itemDir,
-      readmePath: candidate.readmePath,
-    },
-    (data) => `${(data as { absPath: string }).absPath}\n`,
-  );
-}
-
-async function searchProjects(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const query = args.positional.join(' ');
-  if (!query) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects search <query>');
-  const limit = parseIntFlag(args.flags.limit, 50);
-  const statusFilter = parseCsvFlag(args.flags.status);
-  const kindFilter = parseCsvFlag(args.flags.kind);
-  for (const k of ['limit', 'status', 'kind']) delete args.flags[k];
-  assertNoExtraFlags(args);
-
-  const results = await searchAll(conceptionPath, query);
-  const projectHits = results.hits.filter((h) => h.source === 'project');
-
-  // Annotate hits with the matched item's header so the skill can triage.
-  const enriched: (SearchHit & {
-    headerKind?: string;
-    headerStatus?: string;
-    headerApps?: string[];
-  })[] = [];
-  const headerCache = new Map<string, HeaderFields>();
-  for (const hit of projectHits) {
-    if (enriched.length >= limit) break;
-    const projectPath = hit.projectPath;
-    if (!projectPath) {
-      enriched.push(hit);
-      continue;
-    }
-    let header = headerCache.get(projectPath);
-    if (!header) {
-      header = await readHeader(join(projectPath, 'README.md')).catch(() => null as never);
-      if (header) headerCache.set(projectPath, header);
-    }
-    if (!header) {
-      enriched.push(hit);
-      continue;
-    }
-    if (statusFilter && (!header.status || !statusFilter.includes(header.status))) continue;
-    if (kindFilter && (!header.kind || !kindFilter.includes(header.kind))) continue;
-    enriched.push({
-      ...hit,
-      headerKind: header.kind ?? undefined,
-      headerStatus: header.status ?? undefined,
-      headerApps: header.apps,
-    });
-  }
-
-  emit(
-    ctx,
-    {
-      query,
-      hits: enriched,
-      totalBeforeFilter: projectHits.length,
-      truncated: results.truncated,
-    },
-    (data) => formatSearchHuman(data as { hits: typeof enriched; query: string }, conceptionPath),
-    [],
-    { streamField: 'hits' },
-  );
-}
-
-function formatSearchHuman(
-  data: { hits: SearchHit[]; query: string },
-  _conceptionPath: string,
-): string {
-  if (data.hits.length === 0) return `(no project matches for "${data.query}")\n`;
-  const lines: string[] = [];
-  for (const hit of data.hits) {
-    const snippet = hit.snippets[0]?.text.replace(/\s+/g, ' ').slice(0, 120) ?? '';
-    lines.push(`${hit.relPath}: ${snippet}`);
-  }
-  return lines.join('\n') + '\n';
-}
-
-async function validateCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const all = args.flags.all === true;
-  const explicitPath = typeof args.flags.path === 'string' ? args.flags.path : null;
-  const slug = args.positional[0];
-  for (const k of ['all', 'path']) delete args.flags[k];
-  assertNoExtraFlags(args);
-
-  let readmes: string[];
-  if (all) {
-    readmes = await findProjectReadmes(conceptionPath);
-  } else if (explicitPath) {
-    // Path-traversal guard: --path is the one validate input where the
-    // skill/user passes an arbitrary filesystem path, so we resolve and
-    // require it to live under <conception>/projects/. Without this, a
-    // skill bug or shell typo could point validate at an arbitrary
-    // README on disk and we'd dutifully open it.
-    const resolved = resolve(conceptionPath, explicitPath);
-    const projectsRoot = resolve(conceptionPath, 'projects') + sep;
-    if (!resolved.startsWith(projectsRoot)) {
-      throw new CliError(
-        ExitCodes.USAGE,
-        `--path must point inside <conception>/projects/ (got ${explicitPath})`,
-      );
-    }
-    readmes = [resolved];
-  } else if (slug) {
-    const candidate = await resolveSlug(conceptionPath, slug);
-    readmes = [candidate.readmePath];
-  } else {
-    throw new CliError(
-      ExitCodes.USAGE,
-      'Usage: condash-cli projects validate <slug> | --all | --path <readme-path>',
-    );
-  }
-
-  const reports: {
-    path: string;
-    relPath: string;
-    errors: { field: string; message: string }[];
-    warnings: { field: string; message: string }[];
-  }[] = [];
-  let totalErrors = 0;
-
-  for (const readme of readmes) {
-    const raw = await fs.readFile(readme, 'utf8');
-    const fields = parseHeader(raw);
-    const v = validateHeader(fields, readme);
-    const b = validateBody(raw);
-    totalErrors += v.errors.length + b.errors.length;
-    reports.push({
-      path: readme,
-      relPath: relative(conceptionPath, readme),
-      errors: [...v.errors, ...b.errors],
-      warnings: [...v.warnings, ...b.warnings],
-    });
-  }
-
-  // Single envelope discipline: on errors, throw CliError so the dispatcher
-  // emits ONE failure envelope carrying the reports — without this branch,
-  // the success-path emit() would write a success envelope first and
-  // consumers would see two JSON objects on a failed run (the previous bug
-  // the v2.10.17 review caught).
-  if (totalErrors > 0) {
-    throw new CliError(ExitCodes.VALIDATION, `${totalErrors} validation error(s)`, {
-      reports: reports.filter((r) => r.errors.length > 0),
-      totalChecked: reports.length,
-    });
-  }
-
-  emit(
-    ctx,
-    { reports, totalErrors, totalChecked: reports.length },
-    (data) => {
-      const d = data as { reports: typeof reports };
-      const lines: string[] = [];
-      for (const r of d.reports) {
-        if (r.errors.length === 0 && r.warnings.length === 0) continue;
-        lines.push(r.relPath);
-        for (const e of r.errors) lines.push(`  ERROR  ${e.field}: ${e.message}`);
-        for (const w of r.warnings) lines.push(`  warn   ${w.field}: ${w.message}`);
-      }
-      if (lines.length === 0) {
-        return `OK (${d.reports.length} README${d.reports.length === 1 ? '' : 's'} checked)\n`;
-      }
-      return lines.join('\n') + '\n';
-    },
-    [],
-    { streamField: 'reports' },
-  );
-}
-
-async function statusCommand(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const sub = args.positional[0];
-  if (sub === 'get') {
-    const slug = args.positional[1];
-    if (!slug) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects status get <slug>');
-    assertNoExtraFlags(args);
-    const candidate = await resolveSlug(conceptionPath, slug);
-    const header = await readHeader(candidate.readmePath);
-    emit(
-      ctx,
-      { slug: candidate.slug, status: header.status },
-      (d) => `${(d as { status: string | null }).status ?? '(missing)'}\n`,
-    );
-    return;
-  }
-  if (sub === 'set') {
-    const slug = args.positional[1];
-    const value = args.positional[2];
-    if (!slug || !value) {
-      throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects status set <slug> <status>');
-    }
-    if (!(KNOWN_STATUSES as readonly string[]).includes(value)) {
-      validation(`Status '${value}' not in {${KNOWN_STATUSES.join(', ')}}`);
-    }
-    const summary = typeof args.flags.summary === 'string' ? args.flags.summary.trim() : undefined;
-    delete args.flags.summary;
-    assertNoExtraFlags(args);
-    const candidate = await resolveSlug(conceptionPath, slug);
-    const transition = await transitionStatus(candidate.readmePath, value, { summary });
-    const dirtyMarker = await touchDirtyMarker(conceptionPath, 'projects');
-    emit(
-      ctx,
-      {
-        slug: candidate.slug,
-        path: candidate.readmePath,
-        previousStatus: transition.previousStatus,
-        newStatus: transition.newStatus,
-        timelineAppended: transition.timelineAppended,
-        dirtyMarkerTouched: dirtyMarker,
-      },
-      (d) => {
-        const data = d as { previousStatus: string | null; newStatus: string };
-        return `${data.previousStatus ?? '(none)'} → ${data.newStatus}\n`;
-      },
-    );
-    return;
-  }
-  throw new CliError(
-    ExitCodes.USAGE,
-    'Usage: condash-cli projects status <get|set> <slug> [<value>]',
-  );
-}
-
-async function closeProject(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const slug = args.positional[0];
-  if (!slug) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects close <slug>');
-  const newStatus = (args.flags.status as string | undefined) ?? 'done';
-  if (!(KNOWN_STATUSES as readonly string[]).includes(newStatus)) {
-    validation(`Status '${newStatus}' not in {${KNOWN_STATUSES.join(', ')}}`);
-  }
-  const summary = (args.flags.summary as string | undefined)?.trim();
-  const noTouchDirty = args.flags['no-touch-dirty'] === true;
-  for (const k of ['status', 'summary', 'no-touch-dirty']) delete args.flags[k];
-  assertNoExtraFlags(args);
-
-  const candidate = await resolveSlug(conceptionPath, slug);
-  const header = await readHeader(candidate.readmePath);
-  const transition = await transitionStatus(candidate.readmePath, newStatus, { summary });
-
-  const dirtyMarker = noTouchDirty ? false : await touchDirtyMarker(conceptionPath, 'projects');
-
-  const warnings = await leftoverBranchWarnings(conceptionPath, header.branch);
-
-  emit(
-    ctx,
-    {
-      slug: candidate.slug,
-      path: candidate.readmePath,
-      previousStatus: transition.previousStatus,
-      newStatus: transition.newStatus,
-      timelineAppended: transition.timelineAppended,
-      dirtyMarkerTouched: dirtyMarker,
-    },
-    (d) => {
-      const data = d as { slug: string; previousStatus: string | null; newStatus: string };
-      return `Closed ${data.slug}: ${data.previousStatus ?? '(none)'} → ${data.newStatus}\n`;
-    },
-    warnings,
-  );
-}
-
-async function reopenProject(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const slug = args.positional[0];
-  if (!slug) throw new CliError(ExitCodes.USAGE, 'Usage: condash-cli projects reopen <slug>');
-  const target = (args.flags.status as string | undefined) ?? 'now';
-  if (!(KNOWN_STATUSES as readonly string[]).includes(target)) {
-    validation(`Status '${target}' not in {${KNOWN_STATUSES.join(', ')}}`);
-  }
-  if (target === 'done') {
-    validation(`reopen target cannot be 'done' — use \`condash-cli projects close\` instead`);
-  }
-  delete args.flags.status;
-  assertNoExtraFlags(args);
-  const candidate = await resolveSlug(conceptionPath, slug);
-  const header = await readHeader(candidate.readmePath);
-  const previous = (header.status ?? '').toLowerCase();
-  if (previous !== 'done') {
-    throw new CliError(
-      ExitCodes.VALIDATION,
-      `Cannot reopen ${candidate.slug}: previous status is '${previous || '(none)'}', expected 'done'`,
-    );
-  }
-  const transition = await transitionStatus(candidate.readmePath, target);
-  const dirtyMarker = await touchDirtyMarker(conceptionPath, 'projects');
-  emit(
-    ctx,
-    {
-      slug: candidate.slug,
-      path: candidate.readmePath,
-      previousStatus: transition.previousStatus,
-      newStatus: transition.newStatus,
-      timelineAppended: transition.timelineAppended,
-      dirtyMarkerTouched: dirtyMarker,
-    },
-    (d) => {
-      const data = d as { slug: string; previousStatus: string | null; newStatus: string };
-      return `Reopened ${data.slug}: ${data.previousStatus ?? '(none)'} → ${data.newStatus}\n`;
-    },
-  );
-}
-
-interface BackfillEntry {
-  slug: string;
-  readme: string;
-  date: string;
-  source: 'git' | 'mtime';
-  appended: boolean;
-}
-
-/**
- * One-shot migration: walk every `Status: done` README; for any one that has
- * no `- <date> — Closed.` line in its `## Timeline`, append one with a date
- * derived from `git log -1` on the README (cheap heuristic — last touch
- * usually corresponds to the close), falling back to file mtime when git
- * fails (untracked file, no repo, etc.). Run with `--dry-run` to preview.
- *
- * The appended line carries a trailing `(backfill)` marker so backfilled
- * entries are distinguishable from organic close entries.
- */
-async function backfillClosed(
-  args: ParsedArgs,
-  ctx: OutputContext,
-  conceptionPath: string,
-): Promise<void> {
-  const dryRun = args.flags['dry-run'] === true;
-  delete args.flags['dry-run'];
-  assertNoExtraFlags(args);
-  const readmes = await findProjectReadmes(conceptionPath);
-  const candidates: BackfillEntry[] = [];
-  const skipped: { slug: string; reason: string }[] = [];
-  for (const readme of readmes) {
-    const raw = await fs.readFile(readme, 'utf8');
-    const header = parseHeader(raw);
-    if ((header.status ?? '').toLowerCase() !== 'done') continue;
-    const entries = parseTimelineEntries(raw);
-    if (entries.some((e) => /^Closed\b/.test(e.text))) {
-      skipped.push({ slug: basenameOf(readme), reason: 'already has Closed entry' });
-      continue;
-    }
-    let date: string | null = null;
-    let source: 'git' | 'mtime' = 'mtime';
-    try {
-      const { stdout } = await exec(
-        'git',
-        ['log', '-1', '--format=%ad', '--date=short', '--', readme],
-        { cwd: dirname(readme) },
-      );
-      const trimmed = stdout.trim();
-      if (/^\d{4}-\d{2}-\d{2}$/.test(trimmed)) {
-        date = trimmed;
-        source = 'git';
-      }
-    } catch {
-      // git log failed — fall through to mtime.
-    }
-    if (!date) {
-      try {
-        const stat = await fs.stat(readme);
-        date = stat.mtime.toISOString().slice(0, 10);
-      } catch {
-        skipped.push({ slug: basenameOf(readme), reason: 'no date source' });
-        continue;
-      }
-    }
-    candidates.push({
-      slug: basenameOf(readme),
-      readme,
-      date,
-      source,
-      appended: false,
-    });
-  }
-
-  if (!dryRun) {
-    for (const c of candidates) {
-      const line = `- ${c.date} — Closed. (backfill)`;
-      await appendTimelineEntry(c.readme, line);
-      c.appended = true;
-    }
-    if (candidates.length > 0) await touchDirtyMarker(conceptionPath, 'projects');
-  }
-
-  emit(
-    ctx,
-    {
-      dryRun,
-      candidates,
-      skipped,
-      totalScanned: readmes.length,
-    },
-    (data) => {
-      const d = data as { dryRun: boolean; candidates: BackfillEntry[] };
-      const lines: string[] = [];
-      if (d.candidates.length === 0) {
-        lines.push(
-          d.dryRun ? '(no backfill candidates)' : '(no backfill candidates — nothing written)',
-        );
-      } else {
-        lines.push(
-          d.dryRun
-            ? `Would append (${d.candidates.length}):`
-            : `Appended (${d.candidates.length}):`,
-        );
-        for (const c of d.candidates) {
-          lines.push(`  ${c.slug}: ${c.date} (${c.source})`);
-        }
-      }
-      return lines.join('\n') + '\n';
-    },
-  );
-}
-
-function basenameOf(readmePath: string): string {
-  const dir = dirname(readmePath);
-  const parts = dir.split(/[\\/]/);
-  return parts[parts.length - 1] ?? '';
-}
-
-/**
- * Probe the closed item's branch (when the header carries one) and surface
- * a warning if the on-disk worktree or the local branch still exists. Closing
- * an item only flips Status — the actual cleanup verbs are
- * `condash-cli worktrees remove <branch>` and `git branch -d <branch>`, and a
- * silent close lets the miss go unnoticed (this exact thing happened during
- * the parent simplify batch, May 1).
- */
-async function leftoverBranchWarnings(
-  conceptionPath: string,
-  branch: string | null,
-): Promise<string[]> {
-  if (!branch) return [];
-  let state;
-  try {
-    state = await checkBranchState(conceptionPath, branch);
-  } catch {
-    // checkBranchState reads condash.json + queries each repo; if the
-    // probe itself fails we'd rather close cleanly than crash the verb.
-    return [];
-  }
-  const lingeringWorktrees = state.repos.filter((r) => r.worktreeExists);
-  const lingeringBranches = state.repos.filter((r) => r.localBranchExists);
-  if (lingeringWorktrees.length === 0 && lingeringBranches.length === 0) return [];
-
-  const parts: string[] = [];
-  if (lingeringWorktrees.length > 0) {
-    const paths = lingeringWorktrees.map((r) => r.expectedWorktree).join(', ');
-    parts.push(`worktree(s) still on disk at ${paths}`);
-  }
-  if (lingeringBranches.length > 0) {
-    const repos = lingeringBranches.map((r) => r.name).join(', ');
-    parts.push(`local branch '${branch}' still exists in ${repos}`);
-  }
-  return [
-    `${parts.join('; ')} — run \`condash-cli worktrees remove ${branch}\` ` +
-      `then \`git branch -d ${branch}\` to clean up.`,
-  ];
 }
 
 function printSubHelp(): void {
